Fix silly iteration over a sparse matrix
